backend/requirements.txt
Django==4.1.4
django_filter==23.2
djangorestframework==3.14.0
psycopg2-binary==2.9.1